requirements.txt
python-telegram-bot>=20.0
requests
nest_asyncio
flask
